Finish resize table functional
