+++ src/components/table/table.functions.js
@@ -1,0 +1,3 @@
+export function shouldResize(event) {
+  return event.target.dataset.resizer
+}
